qabul.sambhram.uz/requirements.txt
asgiref
Django
django-phonenumber-field
et_xmlfile
numpy
openpyxl
pandas
phonenumbers
pillow
python-dateutil
pytz
six
sqlparse
tzdata

